[FIX] Add conditionals to verify that the retrieved user is an actual user
